Frequency is stored as float in database
Currently, the middle C keyboard is having issues with tartini beats. If you play three tones at once, it sounds like somebody sawing something with power tools.  Fascinating! http://www.animations.physics.unsw.edu.au/jw/beats.htm

+++ db/migrate/20140314155358_create_keys.rb
@@ -2,7 +2,7 @@ class CreateKeys < ActiveRecord::Migration
   def change
     create_table :keys do |t|
       t.integer :keyboard_id
-      t.integer :frequency
+      t.float   :frequency
       t.string  :wave
       t.integer :panner
       t.integer :delay
